+++ src/types.ts
@@ -1,0 +1,8 @@
+import * as vscode from 'vscode';
+
+interface CargoRunnerConfig {
+	prioritySymbolKinds: vscode.SymbolKind[];
+	logLevel: 'debug' | 'info' | 'error';
+}
+
+export { CargoRunnerConfig };
